+++ Dayana/Shared/Persistence/Models/Blog/Commands/CommentCommandsFile.cs
@@ -1,0 +1,190 @@
+﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
+using Dayana.Shared.Infrastructure.Operations;
+using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Dayana.Shared.Persistence.Models.Blog.Commands;
+
+
+#region PostCategoryComments
+
+public class CreatePostCategoryCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public CreatePostCategoryCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostCategoryId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+}
+
+public class DeletePostCategoryCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public DeletePostCategoryCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public int Id { get; set; }
+}
+
+public class UpdatePostCategoryCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public UpdatePostCategoryCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public int Id { get; set; }
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostCategoryId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+}
+
+#endregion
+
+#region PostCategoryIssueComments
+
+public class CreatePostCategoryIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public CreatePostCategoryIssueCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostCategoryIssueId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+}
+
+public class DeletePostCategoryIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public DeletePostCategoryIssueCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public int Id { get; set; }
+}
+
+public class UpdatePostCategoryIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public UpdatePostCategoryIssueCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public int Id { get; set; }
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostCategoryIssueId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+}
+
+
+#endregion
+
+#region PostComment
+
+public class CreatePostCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public CreatePostCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+
+    public RequestInfo RequestInfo { get; set; }
+}
+
+public class DeletePostCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public DeletePostCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public int Id { get; set; }
+
+    public RequestInfo RequestInfo { get; set; }
+}
+
+public class UpdatePostCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public UpdatePostCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+
+    public int Id { get; set; }
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+
+    public RequestInfo RequestInfo { get; set; }
+}
+
+#endregion
+
+#region PostIssueComments
+
+public class CreatePostIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public CreatePostIssueCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostIssueId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+}
+
+public class DeletePostIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public DeletePostIssueCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public int Id { get; set; }
+}
+
+public class UpdatePostIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public UpdatePostIssueCommentCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public RequestInfo RequestInfo { get; set; }
+    public int Id { get; set; }
+    public string CommentText { get; set; }
+    public bool IsReply { get; set; }
+    public int PostIssueId { get; set; }
+    public int CommentOwnerId { get; set; }
+    public int? ReplyToCommentId { get; set; }
+}
+
+#endregion
