+++ Dayana/Shared/Persistence/Models/Blog/Commands/Comments/PostCategoryComments/CreatePostCategoryCommentCommand.cs
@@ -4,19 +4,6 @@ using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.Comments.PostCategoryComments;
-public class CreatePostCategoryCommentCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public CreatePostCategoryCommentCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public RequestInfo RequestInfo { get; set; }
-    public string CommentText { get; set; }
-    public bool IsReply { get; set; }
-    public int PostCategoryId { get; set; }
-    public int CommentOwnerId { get; set; }
-    public int? ReplyToCommentId { get; set; }
-}
 
 
 public class CreatePostCategoryCommentCommandValidator : AbstractValidator<CreatePostCategoryCommentCommand>
