+++ Dayana/Shared/Persistence/Models/Blog/Commands/Comments/PostCategoryComments/UpdatePostCategoryCommentCommand.cs
@@ -4,20 +4,7 @@ using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.Comments.PostCategoryComments;
-public class UpdatePostCategoryCommentCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public UpdatePostCategoryCommentCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public RequestInfo RequestInfo { get; set; }
-    public int Id { get; set; }
-    public string CommentText { get; set; }
-    public bool IsReply { get; set; }
-    public int PostCategoryId { get; set; }
-    public int CommentOwnerId { get; set; }
-    public int? ReplyToCommentId { get; set; }
-}
+
 
 public class UpdatePostCategoryCommentCommandValidator : AbstractValidator<UpdatePostCategoryCommentCommand>
 {
