+++ Dayana/Shared/Persistence/Models/Blog/Commands/Comments/PostCategoryIssueComments/CreatePostCategoryIssueCommentCommand.cs
@@ -5,20 +5,7 @@ using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.Comments.PostCategoryIssueComments;
-public class CreatePostCategoryIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public CreatePostCategoryIssueCommentCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public RequestInfo RequestInfo { get; set; }
 
-    public string CommentText { get; set; }
-    public bool IsReply { get; set; }
-    public int PostCategoryIssueId { get; set; }
-    public int CommentOwnerId { get; set; }
-    public int? ReplyToCommentId { get; set; }
-}
 
 public class UpdatePostCategoryCommentCommandValidatorValidator : AbstractValidator<UpdatePostCategoryCommentCommandValidator>
 {
