+++ Dayana/Shared/Persistence/Models/Blog/Commands/Comments/PostCategoryIssueComments/UpdatePostCategoryIssueCommentCommand.cs
@@ -4,20 +4,6 @@ using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.Comments.PostCategoryIssueComments;
-public class UpdatePostCategoryIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public UpdatePostCategoryIssueCommentCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public RequestInfo RequestInfo { get; set; }
-    public int Id { get; set; }
-    public string CommentText { get; set; }
-    public bool IsReply { get; set; }
-    public int PostCategoryIssueId { get; set; }
-    public int CommentOwnerId { get; set; }
-    public int? ReplyToCommentId { get; set; }
-}
 
 public class UpdatePostCategoryIssueCommentCommandValidator : AbstractValidator<UpdatePostCategoryIssueCommentCommand>
 {
