+++ Dayana/Shared/Persistence/Models/Blog/Commands/Comments/PostComment/CreatePostCommentCommand.cs
@@ -4,22 +4,7 @@ using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.Comments.PostComment;
-public class CreatePostCommentCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public CreatePostCommentCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
 
-    public string CommentText { get; set; }
-    public bool IsReply { get; set; }
-    public int PostId { get; set; }
-    public int CommentOwnerId { get; set; }
-    public int? ReplyToCommentId { get; set; }
-
-    public RequestInfo RequestInfo { get; set; }
-
-}
 
 public class CreatePostCommentCommandValidator : AbstractValidator<CreatePostCommentCommand>
 {
