+++ Dayana/Shared/Persistence/Models/Blog/Commands/Comments/PostIssueComments/CreatePostIssueCommentCommand.cs
@@ -4,19 +4,7 @@ using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.Comments.PostIssueComments;
-public class CreatePostIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public CreatePostIssueCommentCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public RequestInfo RequestInfo { get; set; }
-    public string CommentText { get; set; }
-    public bool IsReply { get; set; }
-    public int PostIssueId { get; set; }
-    public int CommentOwnerId { get; set; }
-    public int? ReplyToCommentId { get; set; }
-}
+
 
 
 public class CreatePostIssueCommentCommandValidator : AbstractValidator<CreatePostIssueCommentCommand>
