+++ Dayana/Shared/Persistence/Models/Blog/Commands/Comments/PostIssueComments/UpdatePostIssueCommentCommand.cs
@@ -4,20 +4,7 @@ using FluentValidation;
 using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.Comments.PostIssueComments;
-public class UpdatePostIssueCommentCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public UpdatePostIssueCommentCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public RequestInfo RequestInfo { get; set; }
-    public int Id { get; set; }
-    public string CommentText { get; set; }
-    public bool IsReply { get; set; }
-    public int PostIssueId { get; set; }
-    public int CommentOwnerId { get; set; }
-    public int? ReplyToCommentId { get; set; }
-}
+
 
 
 public class UpdatePostIssueCommentCommandValidator : AbstractValidator<UpdatePostIssueCommentCommand>
